put needed feature from crm runx module into webshop_2 module
