+++ runex_webshop_2_0/__openerp__.py
@@ -6,27 +6,30 @@
 #
 ##############################################################################
 {
- "name": "RUNEX webshop 2.0",
- "version": "1.0",
- "author": "Linserv AB",
- "category": "Website",
- "summary": "RUNEX webshop 2.0",
- "website": "www.linserv.se",
- "contributors": [
-      'Gediminas Venclova <gediminasv@live.com>'
- ],
- "license": "",
- "depends": ['website', 'website_sale'],
- 'description': """
+    "name": "RUNEX webshop 2.0",
+    "version": "1.0",
+    "author": "Linserv AB",
+    "category": "Website",
+    "summary": "RUNEX webshop 2.0",
+    "website": "www.linserv.se",
+    "contributors": [
+        'Gediminas Venclova <gediminasv@live.com>'
+    ],
+    "license": "",
+    "depends": ['website', 'website_sale'],
+    'description': """
 
         Package of Runex Webshop modifications
 
     """,
- "demo": [],
- "data": ['views/website_sale_template_inherited.xml'],
- "test": [],
- "js": [],
- "css": [],
- "installable": True,
- "auto_install": False,
+    "demo": [],
+    "data": [
+        'views/views.xml',
+        'views/website_sale_template.xml'
+    ],
+    "test": [],
+    "js": [],
+    "css": [],
+    "installable": True,
+    "auto_install": False,
 }
