+++ runex_webshop_2_0/static/src/css/website_sale.css
@@ -1,0 +1,14 @@
+﻿/*=============================================================
+                    Runex
+ ============================================================ */
+
+.ribbon-runex {
+    background-color: #a22a22;
+}
+
+[name="price"]>del.text-danger{
+    display: none;
+}
+.product_price .text-danger{
+    display: none;
+}
