+++ runex_webshop_2_0/views/views.xml
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+    <data noupdate="1">
+        <record model="ir.config_parameter" id="param_search_fields">
+            <field name="key">alt.products.search.fields</field>
+            <field name="value">name description description_sale product_variant_ids.default_code product_variant_ids.default_code attribute_line_ids.value_ids.name public_categ_ids.name website_style_ids.name</field>
+        </record>
+    </data>
+    <data>
+        <record id="product_public_category_form_view" model="ir.ui.view">
+            <field name="name">product.public.category.form.inherited.website_crm_campaign_runex</field>
+            <field name="model">product.public.category</field>
+            <field name="inherit_id" ref="website_sale.product_public_category_form_view"/>
+            <field name="arch" type="xml">
+                <field name="sequence" position="after">
+                    <field name="description"/>
+                </field>
+            </field>
+        </record>
+        <record id="res_lang_form" model="ir.ui.view">
+            <field name="name">res.lang.form.inherited.website_crm_campaign_runex</field>
+            <field name="model">res.lang</field>
+            <field name="inherit_id" ref="base.res_lang_form"/>
+            <field name="arch" type="xml">
+                <field name="translatable" position="after">
+                    <field name="pricelist"/>
+                </field>
+            </field>
+        </record>
+
+
+
+
+
+    </data>
+</openerp>
