--- runex_webshop_2_0/views/website_sale_template_inherited.xml
@@ -1,52 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-
-        <template id="submenu_fix" inherit_id="website.submenu" name="Submenu_inherited">
-            <xpath expr="//li/a[@t-ignore='true']" position="replace">
-                <a t-att-href="(website.menu_id.child_id[0] == submenu) and '/' or submenu.url" t-ignore="true"
-                   t-att-target="'_blank' if submenu.new_window else None">
-                    <t t-if="submenu.name=='Webshop'">
-                        <i class="fa fa-shopping-cart fa-2x"></i>
-                    </t>
-                    <t t-if="not submenu.name=='Webshop'">
-                        <span t-field="submenu.name"/>
-                    </t>
-                </a>
-            </xpath>
-        </template>
-
-        <template id="checkout_fix" inherit_id="website_sale.checkout" name="Checkout_inherited">
-            <xpath expr="//div[@class='col-md-8 oe_cart']/div[3]" position="replace">
-                <div class="clearfix">
-                    <a href="/shop/cart" class="btn btn-default btn-success mb32">
-                        <span class="fa fa-long-arrow-left"/>
-                        Return to Cart
-                    </a>
-                    <a class="btn btn-default btn-success pull-right mb32 a-submit">To checkout
-                        <span class="fa fa-long-arrow-right"/>
-                    </a>
-                </div>
-            </xpath>
-        </template>
-
-        <template id="website_sale.continue_shopping" inherit_id="website_sale.cart" customize_show="True" name="Continue Shopping Button Inherited">
-            <xpath expr="//a[@href='/shop/checkout']" position="before">
-                <a href="/shop" class="btn btn-success mb32"><span class="fa fa-long-arrow-left"/> <span class="hidden-xs">Continue Shopping</span><span class="visible-xs-inline">Continue</span></a>
-            </xpath>
-        </template>
-
-        <template id="products_add_to_cart_inherit" inherit_id="website_sale.products_add_to_cart">
-            <xpath expr="//div[@class='product_price']/t[1]/a" position="attributes">
-                <attribute name="class">btn btn-success btn-s fa fa-shopping-cart a-submit fa-2x</attribute>
-                <attribute name="style">padding: 1px 3px;​</attribute>
-            </xpath>
-            <xpath expr="//div[@class='product_price']/t[2]/a" position="attributes">
-                <attribute name="class">btn btn-success btn-s fa fa-shopping-cart fa-2x</attribute>
-                <attribute name="style">padding: 1px 3px;​</attribute>
-            </xpath>
-        </template>
-
-
-    </data>
-</openerp>
