+++ website_crm_campaign_runex/campaign.py
@@ -113,82 +113,82 @@ class crm_tracking_campaign(models.Model):
         return super(crm_tracking_campaign, self).get_campaigns().filtered(lambda c: (c.reseller_pricelist or c.pricelist) and self.env.context.get('lang') == c.lang.code)
 
 
-class res_currency(models.Model):
-    _inherit = 'res.currency'
+# class res_currency(models.Model):
+#     _inherit = 'res.currency'
+#
+#     country_ids = fields.One2many(comodel_name='res.country', inverse_name='currency_id', string='Countries')
 
-    country_ids = fields.One2many(comodel_name='res.country', inverse_name='currency_id', string='Countries')
 
-
-class res_partner(osv.osv):
-    _inherit = 'res.partner'
-
-    def _property_product_pricelist(self, cr, uid, ids, name, arg, context=None):
-        res = {}
-        for id in ids:
-            # if id == self.pool.get('ir.model.data').get_object_reference(cr, uid, 'base', 'public_partner')[1]:
-            if True:
-                lang = request.context.get('lang')
-                pricelist = self.pool.get('product.pricelist').browse(cr, uid,
-                    self.pool.get('product.pricelist').search(cr, uid,
-                        [('language_ids.code', '=', lang)], context=context), context=context)
-                if not pricelist:
-                    # Fallback if no pricelist found
-                    company_list = self.pool.get('website').search(cr, uid, [], limit=1)
-                    company = self.pool.get('website').browse(cr, uid, company_list).company_id
-                    langs = lang.split('_')
-                    country_code = langs and len(langs) == 2 and langs[1]
-                    companies = (False,) if not company else (False, company.id)
-                    if country_code:
-                        pricelist_list = self.pool.get('product.pricelist').search(
-                            cr,uid,[
-                                ('currency_id.country_ids.code','in',langs),
-                                ('company_id','in',companies),
-                                ('type', '=', 'sale'),
-                                ('version_id', '!=', False),
-                            ],
-                            context=context,limit=1)
-                        if pricelist_list:
-                            pricelist = self.pool.get('product.pricelist').browse(cr, uid,pricelist_list ,context=context)
-                    if not pricelist:
-                        raise Warning(_("No pricelist found for your language! Please contact the administrator."))
-            # else:
-            #     partner = self.pool.get('res.partner').read(cr, uid, id, ['partner_product_pricelist', 'lang', 'commercial_partner_id'], context=context)
-            #     # The compute breaks the commercial fields handling. Check if this partner is it's own commercial partner to account for that.
-            #     if partner['commercial_partner_id'] and partner['commercial_partner_id'][0] != id:
-            #         # Get the pricelist from the commercial partner and move along.
-            #         res[id] = self._property_product_pricelist(cr, uid, [partner['commercial_partner_id'][0]], name, arg, context)[partner['commercial_partner_id'][0]]
-            #         continue
-            #     lang = partner['lang']
-            #     pricelist = self.pool.get('product.pricelist').browse(cr, uid,
-            #         partner['partner_product_pricelist'] and partner['partner_product_pricelist'][0] or [], context=context)
-            if pricelist:
-                #if pricelist.is_fixed:
-                #    res[id] = pricelist.id
-                #else:
-                    # Fetch the active campaign for this partner's language.
-                    c_context = dict(context)
-                    c_context['lang'] = lang
-                    current_campaign = self.pool.get('crm.tracking.campaign').get_campaigns(cr, uid, context=c_context)
-                    if len(current_campaign) > 0:
-                        if pricelist.is_reseller:
-                            res[id] = current_campaign[0].reseller_pricelist.id if current_campaign[0].reseller_pricelist else current_campaign[0].pricelist.id
-                        else:
-                            res[id] = current_campaign[0].pricelist.id if current_campaign[0].pricelist else pricelist.id
-                    else:
-                        res[id] = pricelist.id
-            else:
-                res[id] = False
-        return res
-
-    _columns = {
-        'property_product_pricelist': osv_fields.function(
-            _property_product_pricelist,
-            type='many2one',
-            relation='product.pricelist',
-            domain=[('type','=','sale')],
-            string="Sale Pricelist",
-            help="This pricelist will be used, instead of the default one, for sales to the current partner"),
-    }
+# class res_partner(osv.osv):
+#     _inherit = 'res.partner'
+#
+#     def _property_product_pricelist(self, cr, uid, ids, name, arg, context=None):
+#         res = {}
+#         for id in ids:
+#             # if id == self.pool.get('ir.model.data').get_object_reference(cr, uid, 'base', 'public_partner')[1]:
+#             if True:
+#                 lang = request.context.get('lang')
+#                 pricelist = self.pool.get('product.pricelist').browse(cr, uid,
+#                     self.pool.get('product.pricelist').search(cr, uid,
+#                         [('language_ids.code', '=', lang)], context=context), context=context)
+#                 if not pricelist:
+#                     # Fallback if no pricelist found
+#                     company_list = self.pool.get('website').search(cr, uid, [], limit=1)
+#                     company = self.pool.get('website').browse(cr, uid, company_list).company_id
+#                     langs = lang.split('_')
+#                     country_code = langs and len(langs) == 2 and langs[1]
+#                     companies = (False,) if not company else (False, company.id)
+#                     if country_code:
+#                         pricelist_list = self.pool.get('product.pricelist').search(
+#                             cr,uid,[
+#                                 ('currency_id.country_ids.code','in',langs),
+#                                 ('company_id','in',companies),
+#                                 ('type', '=', 'sale'),
+#                                 ('version_id', '!=', False),
+#                             ],
+#                             context=context,limit=1)
+#                         if pricelist_list:
+#                             pricelist = self.pool.get('product.pricelist').browse(cr, uid,pricelist_list ,context=context)
+#                     if not pricelist:
+#                         raise Warning(_("No pricelist found for your language! Please contact the administrator."))
+#             # else:
+#             #     partner = self.pool.get('res.partner').read(cr, uid, id, ['partner_product_pricelist', 'lang', 'commercial_partner_id'], context=context)
+#             #     # The compute breaks the commercial fields handling. Check if this partner is it's own commercial partner to account for that.
+#             #     if partner['commercial_partner_id'] and partner['commercial_partner_id'][0] != id:
+#             #         # Get the pricelist from the commercial partner and move along.
+#             #         res[id] = self._property_product_pricelist(cr, uid, [partner['commercial_partner_id'][0]], name, arg, context)[partner['commercial_partner_id'][0]]
+#             #         continue
+#             #     lang = partner['lang']
+#             #     pricelist = self.pool.get('product.pricelist').browse(cr, uid,
+#             #         partner['partner_product_pricelist'] and partner['partner_product_pricelist'][0] or [], context=context)
+#             if pricelist:
+#                 #if pricelist.is_fixed:
+#                 #    res[id] = pricelist.id
+#                 #else:
+#                     # Fetch the active campaign for this partner's language.
+#                     c_context = dict(context)
+#                     c_context['lang'] = lang
+#                     current_campaign = self.pool.get('crm.tracking.campaign').get_campaigns(cr, uid, context=c_context)
+#                     if len(current_campaign) > 0:
+#                         if pricelist.is_reseller:
+#                             res[id] = current_campaign[0].reseller_pricelist.id if current_campaign[0].reseller_pricelist else current_campaign[0].pricelist.id
+#                         else:
+#                             res[id] = current_campaign[0].pricelist.id if current_campaign[0].pricelist else pricelist.id
+#                     else:
+#                         res[id] = pricelist.id
+#             else:
+#                 res[id] = False
+#         return res
+#
+#     _columns = {
+#         'property_product_pricelist': osv_fields.function(
+#             _property_product_pricelist,
+#             type='many2one',
+#             relation='product.pricelist',
+#             domain=[('type','=','sale')],
+#             string="Sale Pricelist",
+#             help="This pricelist will be used, instead of the default one, for sales to the current partner"),
+#     }
 
 class ResPartner(models.Model):
     _inherit = 'res.partner'
@@ -206,13 +206,13 @@ class ResPartner(models.Model):
     def search_pricelist(self, operator, value):
         return [('partner_product_pricelist', operator, value)]
 
-class res_lang(models.Model):
-    _inherit = 'res.lang'
-
-    pricelist = fields.Many2one(
-        comodel_name='product.pricelist',
-        domain=[('type', '=', 'sale')],
-        string='Price List')
+# class res_lang(models.Model):
+#     _inherit = 'res.lang'
+#
+#     pricelist = fields.Many2one(
+#         comodel_name='product.pricelist',
+#         domain=[('type', '=', 'sale')],
+#         string='Price List')
 
 
 class product_pricelist(models.Model):
@@ -221,7 +221,7 @@ class product_pricelist(models.Model):
     is_reseller = fields.Boolean(string='Reseller')
     is_fixed = fields.Boolean(string='Fixed')
 
-    language_ids = fields.One2many(comodel_name='res.lang', inverse_name='pricelist', string='Languages')
+    # language_ids = fields.One2many(comodel_name='res.lang', inverse_name='pricelist', string='Languages')
 
 
 class product_template(models.Model):
@@ -244,10 +244,10 @@ class product_product(models.Model):
         self.price = super(product_product, product).with_context({'pricelist': pricelist})._product_price(name,arg)[self.id]
 
 
-class product_public_category(models.Model):
-    _inherit = 'product.public.category'
-
-    description = fields.Text(string='Description')
+# class product_public_category(models.Model):
+#     _inherit = 'product.public.category'
+#
+#     description = fields.Text(string='Description')
     #~ mobile_icon = fields.Char(string='Mobile Icon', help='This icon will display on smaller devices')
 
 class website_campaign(Website):
@@ -270,19 +270,19 @@ class website_sale(website_sale):
     def get_pricelist(self):
         return request.env.user.sudo().partner_id.property_product_pricelist
 
-    def get_user_pricelist(self, user):
-        pricelist = user.partner_id.property_product_pricelist
-        if pricelist.is_fixed:
-            return pricelist
-        else:
-            current_campaign = self.env['crm.tracking.campaign'].get_campaigns()
-            if len(current_campaign) > 0:
-                if pricelist.is_reseller:
-                    return current_campaign[0].reseller_pricelist.id if current_campaign[0].reseller_pricelist else current_campaign[0].pricelist.id
-                else:
-                    return current_campaign[0].pricelist.id if current_campaign[0].pricelist else pricelist
-            else:
-                return pricelist
+    # def get_user_pricelist(self, user):
+    #     pricelist = user.partner_id.property_product_pricelist
+    #     if pricelist.is_fixed:
+    #         return pricelist
+    #     else:
+    #         current_campaign = self.env['crm.tracking.campaign'].get_campaigns()
+    #         if len(current_campaign) > 0:
+    #             if pricelist.is_reseller:
+    #                 return current_campaign[0].reseller_pricelist.id if current_campaign[0].reseller_pricelist else current_campaign[0].pricelist.id
+    #             else:
+    #                 return current_campaign[0].pricelist.id if current_campaign[0].pricelist else pricelist
+    #         else:
+    #             return pricelist
 
 
     @http.route([
@@ -380,142 +380,142 @@ class website_sale(website_sale):
             'url': url,
         })
 
-    @http.route([
-        '/shop',
-        '/shop/page/<int:page>',
-        '/shop/category/<model("product.public.category"):category>',
-        '/shop/category/<model("product.public.category"):category>/page/<int:page>'
-    ], type='http', auth="public", website=True)
-    def shop(self, page=1, category=None, search='', **post):
-        cr, uid, context, pool = request.cr, request.uid, request.context, request.registry
+    # @http.route([
+    #     '/shop',
+    #     '/shop/page/<int:page>',
+    #     '/shop/category/<model("product.public.category"):category>',
+    #     '/shop/category/<model("product.public.category"):category>/page/<int:page>'
+    # ], type='http', auth="public", website=True)
+    # def shop(self, page=1, category=None, search='', **post):
+    #     cr, uid, context, pool = request.cr, request.uid, request.context, request.registry
+    #
+    #     attrib_list = request.httprequest.args.getlist('attrib')
+    #     attrib_values = [map(int, v.split("-")) for v in attrib_list if v]
+    #     attrib_set = set([v[1] for v in attrib_values])
+    #
+    #     domain = self._get_search_domain(search, category, attrib_values)
+    #
+    #     keep = QueryURL('/shop', category=category and int(category), search=search, attrib=attrib_list)
+    #
+    #     #~ if not context.get('pricelist'):
+    #         #~ if request.env['res.users'].browse(request.env.user.id) == request.env.ref('base.public_user'):
+    #             #~ pricelist = request.env['res.lang'].search([('code', '=', request.context.get('lang'))]).pricelist or self.env.ref('product.list0')
+    #         #~ else:
+    #             #~ ppp = user.partner_id.property_product_pricelist
+    #             #~ if ppp:
+    #                 #~ pricelist = get_user_pricelist(request.env.user)
+    #             #~ else:
+    #                 #~ pricelist = request.env['res.lang'].search([('code', '=', request.context.get('lang'))]).pricelist or self.env.ref('product.list0')
+    #         #~ context['pricelist'] = int(pricelist)
+    #     #~ else:
+    #         #~ pricelist = request.env['product.pricelist'].browse(context['pricelist'])
+    #
+    #
+    #     if not context.get('pricelist'):
+    #         pricelist = self.get_pricelist()
+    #         context['pricelist'] = int(pricelist)
+    #     else:
+    #         pricelist = pool.get('product.pricelist').browse(cr, uid, context['pricelist'], context)
+    #
+    #     product_obj = pool.get('product.template')
+    #
+    #     url = "/shop"
+    #     product_count = product_obj.search_count(cr, uid, domain, context=context)
+    #     if search:
+    #         post["search"] = search
+    #     if category:
+    #         category = pool['product.public.category'].browse(cr, uid, int(category), context=context)
+    #         url = "/shop/category/%s" % slug(category)
+    #     if attrib_list:
+    #         post['attrib'] = attrib_list
+    #
+    #     ppg = PPG
+    #     if post.get('limit'):
+    #         limit = post.get('limit')
+    #         try:
+    #             int(limit)
+    #             ppg = abs(int(limit))
+    #         except:
+    #             pass
+    #     pager = request.website.pager(url=url, total=product_count, page=page, step=ppg, scope=7, url_args=post)
+    #     post['order'] = post.get('order', 'name')
+    #     product_ids = product_obj.search(cr, uid, domain, limit=ppg, offset=pager['offset'], order=self._get_search_order(post), context=context)
+    #     products = product_obj.browse(cr, uid, product_ids, context=context)
+    #
+    #     style_obj = pool['product.style']
+    #     style_ids = style_obj.search(cr, uid, [], context=context)
+    #     styles = style_obj.browse(cr, uid, style_ids, context=context)
+    #
+    #     category_obj = pool['product.public.category']
+    #     category_ids = category_obj.search(cr, uid, [('parent_id', '=', False)], context=context)
+    #     categs = category_obj.browse(cr, uid, category_ids, context=context)
+    #
+    #     attributes_obj = request.registry['product.attribute']
+    #     attributes_ids = attributes_obj.search(cr, uid, [], context=context)
+    #     attributes = attributes_obj.browse(cr, uid, attributes_ids, context=context)
+    #
+    #     from_currency = pool.get('product.price.type')._get_field_currency(cr, uid, 'list_price', context)
+    #     to_currency = pricelist.currency_id
+    #     compute_currency = lambda price: pool['res.currency']._compute(cr, uid, from_currency, to_currency, price, context=context)
+    #     view_type = 'grid_view'
+    #     if post.get('view_type') and post.get('view_type') == 'list_view':
+    #         view_type = 'list_view'
+    #
+    #     values = {
+    #         'search': search,
+    #         'category': category,
+    #         'attrib_values': attrib_values,
+    #         'attrib_set': attrib_set,
+    #         'page': page,
+    #         'pager': pager,
+    #         'pricelist': pricelist,
+    #         'products': products,
+    #         'bins': table_compute().process(products, ppg),
+    #         'rows': PPR,
+    #         'styles': styles,
+    #         'categories': categs,
+    #         'attributes': attributes,
+    #         'compute_currency': compute_currency,
+    #         'keep': keep,
+    #         'style_in_product': lambda style, product: style.id in [s.id for s in product.website_style_ids],
+    #         'attrib_encode': lambda attribs: werkzeug.url_encode([('attrib',i) for i in attribs]),
+    #         'product_count': product_count,
+    #         'view_type': view_type,
+    #         'limit': ppg,
+    #         'url': url,
+    #     }
+    #     return request.website.render("website_sale.products", values)
 
-        attrib_list = request.httprequest.args.getlist('attrib')
-        attrib_values = [map(int, v.split("-")) for v in attrib_list if v]
-        attrib_set = set([v[1] for v in attrib_values])
+    # def get_translation(self, product):
+    #     try:
+    #         return request.env['ir.translation'].search([('res_id', '=', product.id), ('name', '=', 'product.template,name'), ('type', '=', 'model'), ('lang', '=', context.get('lang'))])[-1].value
+    #     except:
+    #         return product.name
 
-        domain = self._get_search_domain(search, category, attrib_values)
-
-        keep = QueryURL('/shop', category=category and int(category), search=search, attrib=attrib_list)
-
-        #~ if not context.get('pricelist'):
-            #~ if request.env['res.users'].browse(request.env.user.id) == request.env.ref('base.public_user'):
-                #~ pricelist = request.env['res.lang'].search([('code', '=', request.context.get('lang'))]).pricelist or self.env.ref('product.list0')
-            #~ else:
-                #~ ppp = user.partner_id.property_product_pricelist
-                #~ if ppp:
-                    #~ pricelist = get_user_pricelist(request.env.user)
-                #~ else:
-                    #~ pricelist = request.env['res.lang'].search([('code', '=', request.context.get('lang'))]).pricelist or self.env.ref('product.list0')
-            #~ context['pricelist'] = int(pricelist)
-        #~ else:
-            #~ pricelist = request.env['product.pricelist'].browse(context['pricelist'])
-
-
-        if not context.get('pricelist'):
-            pricelist = self.get_pricelist()
-            context['pricelist'] = int(pricelist)
-        else:
-            pricelist = pool.get('product.pricelist').browse(cr, uid, context['pricelist'], context)
-
-        product_obj = pool.get('product.template')
-
-        url = "/shop"
-        product_count = product_obj.search_count(cr, uid, domain, context=context)
-        if search:
-            post["search"] = search
-        if category:
-            category = pool['product.public.category'].browse(cr, uid, int(category), context=context)
-            url = "/shop/category/%s" % slug(category)
-        if attrib_list:
-            post['attrib'] = attrib_list
-
-        ppg = PPG
-        if post.get('limit'):
-            limit = post.get('limit')
-            try:
-                int(limit)
-                ppg = abs(int(limit))
-            except:
-                pass
-        pager = request.website.pager(url=url, total=product_count, page=page, step=ppg, scope=7, url_args=post)
-        post['order'] = post.get('order', 'name')
-        product_ids = product_obj.search(cr, uid, domain, limit=ppg, offset=pager['offset'], order=self._get_search_order(post), context=context)
-        products = product_obj.browse(cr, uid, product_ids, context=context)
-
-        style_obj = pool['product.style']
-        style_ids = style_obj.search(cr, uid, [], context=context)
-        styles = style_obj.browse(cr, uid, style_ids, context=context)
-
-        category_obj = pool['product.public.category']
-        category_ids = category_obj.search(cr, uid, [('parent_id', '=', False)], context=context)
-        categs = category_obj.browse(cr, uid, category_ids, context=context)
-
-        attributes_obj = request.registry['product.attribute']
-        attributes_ids = attributes_obj.search(cr, uid, [], context=context)
-        attributes = attributes_obj.browse(cr, uid, attributes_ids, context=context)
-
-        from_currency = pool.get('product.price.type')._get_field_currency(cr, uid, 'list_price', context)
-        to_currency = pricelist.currency_id
-        compute_currency = lambda price: pool['res.currency']._compute(cr, uid, from_currency, to_currency, price, context=context)
-        view_type = 'grid_view'
-        if post.get('view_type') and post.get('view_type') == 'list_view':
-            view_type = 'list_view'
-
-        values = {
-            'search': search,
-            'category': category,
-            'attrib_values': attrib_values,
-            'attrib_set': attrib_set,
-            'page': page,
-            'pager': pager,
-            'pricelist': pricelist,
-            'products': products,
-            'bins': table_compute().process(products, ppg),
-            'rows': PPR,
-            'styles': styles,
-            'categories': categs,
-            'attributes': attributes,
-            'compute_currency': compute_currency,
-            'keep': keep,
-            'style_in_product': lambda style, product: style.id in [s.id for s in product.website_style_ids],
-            'attrib_encode': lambda attribs: werkzeug.url_encode([('attrib',i) for i in attribs]),
-            'product_count': product_count,
-            'view_type': view_type,
-            'limit': ppg,
-            'url': url,
-        }
-        return request.website.render("website_sale.products", values)
-
-    def get_translation(self, product):
-        try:
-            return request.env['ir.translation'].search([('res_id', '=', product.id), ('name', '=', 'product.template,name'), ('type', '=', 'model'), ('lang', '=', context.get('lang'))])[-1].value
-        except:
-            return product.name
-
-    def _get_search_domain(self, search, category, attrib_values):
-        domain = request.website.sale_product_domain()
-
-        if search:
-            search_fields = request.env['ir.config_parameter'].get_param('alt.product.search.fields', 'name description description_sale product_variant_ids.default_code').split(" ")
-            for srch in search.split(" "):
-                domain += ['|' for x in range(len(search_fields) - 1)] + [(f, 'ilike', srch) for f in search_fields]
-        if category:
-            domain += [('public_categ_ids', 'child_of', int(category))]
-
-        if attrib_values:
-            attrib = None
-            ids = []
-            for value in attrib_values:
-                if not attrib:
-                    attrib = value[0]
-                    ids.append(value[1])
-                elif value[0] == attrib:
-                    ids.append(value[1])
-                else:
-                    domain += [('attribute_line_ids.value_ids', 'in', ids)]
-                    attrib = value[0]
-                    ids = [value[1]]
-            if attrib:
-                domain += [('attribute_line_ids.value_ids', 'in', ids)]
-
-        return domain
+    # def _get_search_domain(self, search, category, attrib_values):
+    #     domain = request.website.sale_product_domain()
+    #
+    #     if search:
+    #         search_fields = request.env['ir.config_parameter'].get_param('alt.product.search.fields', 'name description description_sale product_variant_ids.default_code').split(" ")
+    #         for srch in search.split(" "):
+    #             domain += ['|' for x in range(len(search_fields) - 1)] + [(f, 'ilike', srch) for f in search_fields]
+    #     if category:
+    #         domain += [('public_categ_ids', 'child_of', int(category))]
+    #
+    #     if attrib_values:
+    #         attrib = None
+    #         ids = []
+    #         for value in attrib_values:
+    #             if not attrib:
+    #                 attrib = value[0]
+    #                 ids.append(value[1])
+    #             elif value[0] == attrib:
+    #                 ids.append(value[1])
+    #             else:
+    #                 domain += [('attribute_line_ids.value_ids', 'in', ids)]
+    #                 attrib = value[0]
+    #                 ids = [value[1]]
+    #         if attrib:
+    #             domain += [('attribute_line_ids.value_ids', 'in', ids)]
+    #
+    #     return domain
